Move list of stations to station module
